// test/integration/test/ports/_fixtures/naming-conflict-builtin-vs-enum/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
type DefaultLiteralValue<CodecId extends string, _Encoded> = CodecId extends keyof CodecTypes
  ? CodecTypes[CodecId]['output']
  : _Encoded;

export type FieldOutputTypes = {
  readonly public: {
    readonly EnumHolderKeys: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
    readonly EnumHolderPromise: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
    readonly EnumHolderResult: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
    readonly EnumHolderUnion: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly EnumHolderKeys: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
    readonly EnumHolderPromise: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
    readonly EnumHolderResult: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
    readonly EnumHolderUnion: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly enum_holder_keys: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
    readonly enum_holder_promise: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
    readonly enum_holder_result: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
    readonly enum_holder_union: { readonly id: Char<24>; readonly value: 'ONE' | 'TWO' };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly enum_holder_keys: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
    readonly enum_holder_promise: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
    readonly enum_holder_result: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
    readonly enum_holder_union: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly value: 'ONE' | 'TWO';
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;